pyproject.toml
[tool.poetry]
name = "power-rangers-api"
version = "0.1.0"
description = "Power Rangers API"
authors = ["Reinier van der Windt <reiniervdwindt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
coverage = "^6.5.0"
drf-spectacular = "^0.24.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
isort = "^5.10.1"
safety = "^2.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
